refactor: move some files around
