+++ apps/game/server/index.ts
@@ -22,6 +22,8 @@ export * from "../shared/client-view-distance-rect";
 export * from "../shared/area/area-id";
 export * from "../shared/area/area-resource";
 export * from "../shared/area/tiled-resource";
-export * from "../user/index";
+export * from "./user/session";
+export * from "./user/auth";
+export * from "./user/define-roles";
 export * from "./rpc";
 export * from "./rpc.slice";
