+++ apps/game/server/npc/router.ts
@@ -1,8 +1,8 @@
 import { ctxGameStateMachine } from "../game-state";
 import { ctxAreaLookup } from "../area/lookup";
-import { defineRoles } from "../../user/define-roles";
 import { rpc } from "../rpc";
-import { roles } from "../../user";
+import { roles } from "../user/auth";
+import { defineRoles } from "../user/define-roles";
 import { ctxNpcService } from "./service";
 import { spawnNpcInstance } from "./npc-spawn-behavior";
