--- apps/game/user/index.ts
@@ -1,3 +1,0 @@
-export * from "./session";
-export * from "./auth";
-export * from "./define-roles";
